The RandomLotkaVolterraCavity Module

RandomLotkaVolterraCavity.RandomLotkaVolterraCavity — Module

Module for "RandomLotkaVolterraCavity.jl" – A Julia package for the study of the fixed-point structure of the Lotka-Volterra model on random graphs with cavity method.

Exports

AbstractNoiseKind

CavityFP

CavityFP_q0

Deterministic

MarginalFP

MarginalFP_q0

Model

ModelDisordered

ModelDisorderedFC

NodeFP

NodeFP_q0

PopFP

PopFP_q0

PopJ

analytic_FC

run_GECaM_FP

run_GECaM_FP_q0

run_MC

run_MC_ODE

sample_couplings

Module Index

RandomLotkaVolterraCavity.RandomLotkaVolterraCavity

RandomLotkaVolterraCavity.AbstractNoiseKind

RandomLotkaVolterraCavity.CavityFP

RandomLotkaVolterraCavity.CavityFP_q0

RandomLotkaVolterraCavity.Deterministic

RandomLotkaVolterraCavity.MarginalFP

RandomLotkaVolterraCavity.MarginalFP_q0

RandomLotkaVolterraCavity.Model

RandomLotkaVolterraCavity.ModelDisordered

RandomLotkaVolterraCavity.ModelDisorderedFC

RandomLotkaVolterraCavity.NodeFP

RandomLotkaVolterraCavity.NodeFP_q0

RandomLotkaVolterraCavity.PopFP

RandomLotkaVolterraCavity.PopFP_q0

RandomLotkaVolterraCavity.PopJ

RandomLotkaVolterraCavity.analytic_FC

RandomLotkaVolterraCavity.run_GECaM_FP

RandomLotkaVolterraCavity.run_GECaM_FP

RandomLotkaVolterraCavity.run_GECaM_FP_q0

RandomLotkaVolterraCavity.run_MC

RandomLotkaVolterraCavity.run_MC

RandomLotkaVolterraCavity.run_MC

RandomLotkaVolterraCavity.run_MC_ODE

RandomLotkaVolterraCavity.run_MC_ODE

RandomLotkaVolterraCavity.run_MC_ODE

RandomLotkaVolterraCavity.sample_couplings

Detailed API

RandomLotkaVolterraCavity.AbstractNoiseKind — Type

AbstractNoiseKind

An abstract type that represents the kind of noise in the model. It can be one of the following:

Deterministic: Represents deterministic dynamics, i.e., no noise.

Demographic: Represents demographic noise, i.e., emerging from birth and death processes.

Environmental: Represents environmental noise, i.e., arising from external factors affecting the population dynamics.

RandomLotkaVolterraCavity.CavityFP — Type

CavityFP{Deterministic, I, RT}

A type that represents a cavity in the fixed-point algorithm for deterministic dynamics.

Fields

i::Integer: Node index

j::Integer: Neighbor index

mu::Real: Cavity mean

q::Real: Cavity correlation

chi::Real: Cavity susceptibility

psi::Real: Cavity survival probability

RandomLotkaVolterraCavity.CavityFP_q0 — Type

CavityFP_q0{Deterministic, I, RT}

A type that represents a cavity in the fixed-point algorithm for deterministic dynamics, using the simplified version where q = 0.

Fields

i::Integer: Node index

j::Integer: Neighbor index

x::Real: Cavity abundance

chi::Real: Cavity susceptibility

RandomLotkaVolterraCavity.Deterministic — Type

Deterministic <: AbstractNoiseKind

A type that represents deterministic dynamics, i.e., no noise.

RandomLotkaVolterraCavity.MarginalFP — Type

MarginalFP{Deterministic, I, RT}

A type that represents a marginal in the fixed-point algorithm for deterministic dynamics.

Fields

i::Integer: Node index

mu::Real: Marginal mean

q::Real: Marginal correlation

chi::Real: Marginal susceptibility

psi::Real: Marginal survival probability

RandomLotkaVolterraCavity.MarginalFP_q0 — Type

MarginalFP_q0{Deterministic, I, RT}

A type that represents a marginal in the fixed-point algorithm for deterministic dynamics, using the simplified version where q = 0.

Fields

i::Integer: Node index

x::Real: Marginal abundance

chi::Real: Marginal susceptibility

RandomLotkaVolterraCavity.Model — Type

Model{NK<:AbstractNoiseKind, I<:Integer, RT<:Real, MT<:AbstractMatrix{RT}}<

A type to store the main parameters for simulating a generalized Lotka-Volterra on a graph.

Fields

N::Integer: Number of nodes in the system.

M::Integer: Number of discretized time steps for the simulation trajectories.

J::AbstractMatrix{RT} where RT<:Real: Adjacency matrix representing the interactions between nodes. Can be a dense or sparse matrix with Real-valued elements.

lam::Real: Immigration rate (lambda).

p0::Distributions.Distribution: Distribution or sampler for initial conditions (p0) of trajectories. This could be a function (e.g., rng -> initial_state), a distribution object from a package like Distributions.jl, or a custom sampler type.

noise::AbstractNoiseKind: Noise of the system.

RandomLotkaVolterraCavity.ModelDisordered — Type

ModelDisordered{NK<:AbstractNoiseKind, I<:Integer, RT<:Real, D1<:Distribution, D2<:Distribution, D3<:Distribution, FT<:Function}

A type to store the main parameters for simulating a generalized Lotka-Volterra system with disordered couplings on a random graph.

Fields

N::Integer: Number of nodes in the system.

K::Real: Average degree of the interaction graph.

M::Integer: Number of discretized time steps for the simulation trajectories.

generate_adj::Function: Function to generate the adjacency matrix

deg_pdf::Distributions.Distribution: Distribution or sampler for the degree of the nodes. This should be a distribution object from the package Distributions.jl.

deg_cav_pdf::Distributions.Distribution: Distribution or sampler for the cavity degree of the nodes. This should be a distribution object from the package Distributions.jl.

m::Real: Average coupling strength

sigma2::Real: Variance of the coupling strength

corr::Real: Correlation of the couplings strength

lam::Real: Immigration rate (lambda).

p0::Distributions.Distribution: Distribution or sampler for initial conditions (p0) of trajectories. This should be a distribution object from the package Distributions.jl.

noise::AbstractNoiseKind: Noise of the system.

RandomLotkaVolterraCavity.ModelDisorderedFC — Type

ModelDisorderedFC{NK<:AbstractNoiseKind, I<:Integer, RT<:Real, D<:Distribution}

A type to store the main parameters for simulating a generalized Lotka-Volterra system with disordered couplings on a fully-connected graph.

Fields

N::Integer: Number of nodes in the system.

M::Integer: Number of discretized time steps for the simulation trajectories.

m::Real: Average coupling strength

sigma2::Real: Variance of the coupling strength

corr::Real: Correlation of the couplings strength

lam::Real: Immigration rate (lambda).

p0::Distributions.Distribution: Distribution or sampler for initial conditions (p0) of trajectories. This should be a distribution object from the package Distributions.jl.

noise::AbstractNoiseKind: Noise of the system.

RandomLotkaVolterraCavity.NodeFP — Type

NodeFP{Deterministic, I, RT}

A type that represents a node in the fixed-point algorithm for deterministic dynamics.

Fields

i::Integer: Node index

neighs::Vector{I} where I<:Integer: Neighbors' indices

neighs_idx::Dict{I, I} where I<:Integer: Dictionary to map neighbors' indices to their positions in the neighs vector

cavs::Array{CavityFP{Deterministic, I, RT}, 1} where {Deterministic, I<:Integer, RT<:Real}: Cavity messages

marg::MarginalFP: Marginal

RandomLotkaVolterraCavity.NodeFP_q0 — Type

CavityFP_q0{Deterministic, I, RT}

A type that represents a cavity in the fixed-point algorithm for deterministic dynamics, using the simplified version where q = 0.

Fields

i::Integer: Node index

neighs::Vector{I} where I<:Integer: Neighbors' indices

neighs_idx::Dict{I, I} where I<:Integer: Dictionary to map neighbors' indices to their positions in the neighs vector

cavs::Array{CavityFP_q0{Deterministic, I, RT}, 1} where {Deterministic, I<:Integer, RT<:Real}: Cavity messages

marg::MarginalFP_q0: Marginal

RandomLotkaVolterraCavity.PopFP — Type

PopFP{Deterministic, I, RT}

A type that represents a population of messages for the fixed-point algorithm in deterministic dynamics.

Fields

mu_pop::Vector{RT} where RT<:Real: Population of averages

q_pop::Vector{RT} where RT<:Real: Population of cavity correlations

chi_pop::Vector{RT} where RT<:Real: Population of cavity susceptibilities

psi_pop::Vector{RT} where RT<:Real: Population of cavity survival probabilities

RandomLotkaVolterraCavity.PopFP_q0 — Type

PopFP_q0{Deterministic, I, RT}

A type that represents a population of messages for the fixed-point algorithm in deterministic dynamics. It uses the simplified version where q = 0.

Fields

x_pop::Vector{RT} where RT<:Real: Population of abundances

chi_pop::Vector{RT} where RT<:Real: Population of susceptibilities

RandomLotkaVolterraCavity.PopJ — Type

PopJ{Deterministic, I, RT}

A type that represents a population of interaction strengths J and J' for a given number of nodes.

Fields

J_pop::Vector{RT} where RT<:Real: Population of interaction strengths J

Jp_pop::Vector{RT} where RT<:Real: Population of interaction strengths J'

RandomLotkaVolterraCavity.run_MC — Method

run_MC(model::Model{NK, I, RT, MT, D}, dt::RT; rng=Xoshiro(1234), showprogress=false, idxs_tsave=nothing, divergence_threshold=1e6) where {NK<:AbstractNoiseKind, I<:Integer, RT<:Real, MT<:AbstractMatrix{RT}, D<:Distribution}

Run a Monte Carlo simulation of the model for a given time step dt. The function samples trajectories of the model and saves them at specified time indices.

Arguments

model::Model{NK, I, RT, MT, D}: The model to simulate, where NK is the noise kind, I is the integer type for indices, RT is the real type, MT is the matrix type, and D is the distribution type.

dt::RT: Time step for the simulation.

Optional arguments

rng::AbstractRNG: Random number generator (default Xoshiro(1234)).

showprogress::Bool: Whether to show a progress bar (default false).

idxs_tsave::Union{Nothing, Vector{Int}}: Indices at which to save the trajectory (default nothing, which saves at every time step).

divergence_threshold::RT: Threshold for divergence detection (default 1e6).

Output

traj::Matrix{RT}: Matrix of sampled trajectories, where each column corresponds to a time point.

tsave::Vector{RT}: Vector of time points at which the trajectories are saved.

convergence::Bool: Boolean indicating whether the integration was successful (true) or diverged/failed (false).

RandomLotkaVolterraCavity.run_MC — Method

run_MC(model_dis::ModelDisorderedFC{NK, I, RT, D}, dt::RT; rng=Xoshiro(1234), showprogress=false, idxs_tsave=nothing, divergence_threshold=1e6) where {NK<:AbstractNoiseKind, I<:Integer, RT<:Real, D<:Distribution}

Run a Monte Carlo simulation of the fully-connected disordered model for a given time step dt by sampling an instance of disorder. The function samples trajectories of the model and saves them at specified time indices.

Arguments

model_dis::ModelDisorderedFC{NK, I, RT, D}: The disordered model to simulate, where NK is the noise kind.

dt::RT: Time step for the simulation.

Optional arguments

rng::AbstractRNG: Random number generator (default Xoshiro(1234)).

showprogress::Bool: Whether to show a progress bar (default false).

idxs_tsave::Union{Nothing, Vector{Int}}: Indices at which to save the trajectory (default nothing, which saves at every time step).

divergence_threshold::RT: Threshold for divergence detection (default 1e6).

Output

traj::Matrix{RT}: Matrix of sampled trajectories, where each column corresponds to a time point.

tsave::Vector{RT}: Vector of time points at which the trajectories are saved.

convergence::Bool: Boolean indicating whether the integration was successful (true) or diverged/failed (false).

RandomLotkaVolterraCavity.run_MC — Method

run_MC(model_dis::ModelDisordered{NK, I, RT, D1, D2, D3, FT}, dt::RT; rng=Xoshiro(1234), showprogress=false, idxs_tsave=nothing, divergence_threshold=1e6) where {NK<:AbstractNoiseKind, I<:Integer, RT<:Real, D1<:Distribution, D2<:Distribution, D3<:Distribution, FT<:Function}

Run a Monte Carlo simulation of the disordered model for a given time step dt by sampling an instance of disorder. The function samples trajectories of the model and saves them at specified time indices.

Arguments

model_dis::ModelDisordered{NK, I, RT, D1, D2, D3, FT}: The disordered model to simulate, where NK is the noise kind.

dt::RT: Time step for the simulation.

Optional arguments

rng::AbstractRNG: Random number generator (default Xoshiro(1234)).

showprogress::Bool: Whether to show a progress bar (default false).

idxs_tsave::Union{Nothing, Vector{Int}}: Indices at which to save the trajectory (default nothing, which saves at every time step).

divergence_threshold::RT: Threshold for divergence detection (default 1e6).

Output

traj::Matrix{RT}: Matrix of sampled trajectories, where each column corresponds to a time point.

tsave::Vector{RT}: Vector of time points at which the trajectories are saved.

convergence::Bool: Boolean indicating whether the integration was successful (true) or diverged/failed (false).

RandomLotkaVolterraCavity.sample_couplings — Method

sample_couplings(m::RT, sigma2::RT, corr::RT, K::RT; rng::AbstractRNG=Xoshiro(1234)) where {RT<:Real}

Sample couplings J and Jp from a bivariate Gaussian distribution with means m/K, variances sigma2/K, and correlation corr.

Arguments

m::RT: Mean value of the couplings.

sigma2::RT: Variance of the couplings.

corr::RT: Correlation between the couplings.

K::RT: Average degree of the network (used for proper scaling of the couplings).

Returns

J::RT: Sampled coupling J.

Jp::RT: Sampled coupling J'.

Example

m = 1.0
sigma2 = 0.1
corr = 0.5

J, J_prime = sample_couplings(m, sigma2, corr, 10)

RandomLotkaVolterraCavity.run_MC_ODE — Method

run_MC_ODE(model::Model{NK, I, RT, MT, D}, dt::RT; rng=Xoshiro(1234), tsave=nothing, divergence_threshold=1e6, stopateq=false, eq_threshold=1e-10, min_t_eq=nothing, verbose=false) where {NK<:AbstractNoiseKind, I<:Integer, RT<:Real, MT<:AbstractMatrix{RT}, D<:Distribution}

Run a Monte Carlo simulation of the model for a given time step dt. It uses the OrdinaryDiffEq.jl package to integrate the ODEs. The function samples trajectories of the model and saves them at specified time indices.

Arguments

model::Model{NK, I, RT, MT, D}: The model to simulate, where NK is the noise kind, I is the integer type for indices, RT is the real type, MT is the matrix type, and D is the distribution type.

dt::RT: Time step for the simulation.

Optional arguments

rng::AbstractRNG: Random number generator (default Xoshiro(1234)).

tsave::Union{Nothing, Vector{RT}}: Vector of time points at which the trajectories are saved.

divergence_threshold::RT: Threshold for divergence detection (default 1e6).

stopateq::Bool: Whether to stop the simulation upon reaching a steady state (default false).

eq_threshold::RT: Threshold for steady state detection (default 1e-10).

min_t_eq::Union{Nothing, RT}: Minimum time before checking for steady state (default nothing, which uses the default in TerminateSteadyState).

verbose::Bool: Whether to print information about the simulation (default false).

Output

traj::Matrix{RT}: Matrix of sampled trajectories, where each column corresponds to a time point.

tsave::Vector{RT}: Vector of time points at which the trajectories are saved.

convergence::Bool: Boolean indicating whether the integration was successful (true) or diverged/failed (false).

t_eq::RT: Equilibriation time. If stopateq is false, returns the final time of integration.

RandomLotkaVolterraCavity.run_MC_ODE — Method

run_MC_ODE(model_dis::ModelDisorderedFC{NK, I, RT, D}, dt::RT; rng=Xoshiro(1234), tsave=nothing, divergence_threshold=1e6, stopateq=false, eq_threshold=1e-10, min_t_eq=1e2, verbose=false) where {NK<:AbstractNoiseKind, I<:Integer, RT<:Real, D<:Distribution}

Run a Monte Carlo simulation of the fully-connected disordered model for a given time step dt, by sampling an instance of disorder. It uses the OrdinaryDiffEq.jl package to integrate the ODEs. The function samples trajectories of the model and saves them at specified time indices.

Arguments

model_dis::ModelDisorderedFC{NK, I, RT, D}: The disordered model to simulate, where NK is the noise kind.

dt::RT: Time step for the simulation.

Optional arguments

rng::AbstractRNG: Random number generator (default Xoshiro(1234)).

tsave::Union{Nothing, Vector{Int}}: Vector of time points at which the trajectories are saved.

divergence_threshold::RT: Threshold for divergence detection (default 1e6).

stopateq::Bool: Whether to stop the simulation upon reaching a steady state (default false).

eq_threshold::RT: Threshold for steady state detection (default 1e-10).

min_t_eq::Union{Nothing, RT}: Minimum time before checking for steady state (default nothing, which uses the default in TerminateSteadyState).

verbose::Bool: Whether to print information about the simulation (default false).

Output

traj::Matrix{RT}: Matrix of sampled trajectories, where each column corresponds to a time point.

tsave::Vector{RT}: Vector of time points at which the trajectories are saved.

convergence::Bool: Boolean indicating whether the integration was successful (true) or diverged/failed (false).

t_eq::RT: Equilibriation time. If stopateq is false, returns the final time of integration.

RandomLotkaVolterraCavity.run_MC_ODE — Method

run_MC_ODE(model_dis::ModelDisordered{NK, I, RT, D1, D2, D3, FT}, dt::RT; rng=Xoshiro(1234), tsave=nothing, divergence_threshold=1e6, stopateq=false, eq_threshold=1e-10, min_t_eq=1e2, verbose=false) where {NK<:AbstractNoiseKind, I<:Integer, RT<:Real, D1<:Distribution, D2<:Distribution, D3<:Distribution, FT<:Function}

Run a Monte Carlo simulation of the disordered model for a given time step dt by sampling an instance of disorder. It uses the OrdinaryDiffEq.jl package to integrate the ODEs. The function samples trajectories of the model and saves them at specified time indices.

Arguments

model_dis::ModelDisordered{NK, I, RT, D1, D2, D3, FT}: The disordered model to simulate, where NK is the noise kind.

dt::RT: Time step for the simulation.

Optional arguments

rng::AbstractRNG: Random number generator (default Xoshiro(1234)).

tsave::Union{Nothing, Vector{Int}}: Vector of time points at which the trajectories are saved.

divergence_threshold::RT: Threshold for divergence detection (default 1e6).

stopateq::Bool: Whether to stop the simulation upon reaching a steady state (default false).

eq_threshold::RT: Threshold for steady state detection (default 1e-10).

min_t_eq::Union{Nothing, RT}: Minimum time before checking for steady state (default nothing, which uses the default in TerminateSteadyState).

verbose::Bool: Whether to print information about the simulation (default false).

Output

traj::Matrix{RT}: Matrix of sampled trajectories, where each column corresponds to a time point.

tsave::Vector{RT}: Vector of time points at which the trajectories are saved.

convergence::Bool: Boolean indicating whether the integration was successful (true) or diverged/failed (false).

t_eq::RT: Equilibriation time. If stopateq is false, returns the final time of integration.

RandomLotkaVolterraCavity.analytic_FC — Method

analytic_FC(m::RT, gamma::RT, npoints::I, Delta_min::RT, Delta_max::RT) where {RT<:Real, I<:Integer}

Returns the analytic solution of the fully-connected system for given parameters.

Arguments

m::RT: Average coupling strength.

gamma::RT: Coupling strength.

npoints::I: Number of points in the range.

Delta_min::RT: Minimum value of the parameter Delta.

Delta_max::RT: Maximum value of the parameter Delta.

Returns

Deltas::Vector{RT}: Values of the parameter Delta.

mus::Vector{RT}: Fixed-point values of the mean abundances mu.

qs::Vector{RT}: Fixed-point values of the mean squared abundances q.

chis::Vector{RT}: Fixed-point values of the susceptibilities chi.

sigma2s::Vector{RT}: Variances sigma2.

phis::Vector{RT}: Fixed-point values of the survival probabilities phi.

RandomLotkaVolterraCavity.run_GECaM_FP — Method

run_GECaM_FP(model::Model{Deterministic, I, RT, MT, D}, max_iter::I, conv_threshold::RT, damp::RT; init_type::Symbol=:random, mu0::RT=0.0, q0::RT=0.0, chi0::RT=0.0, rng::AbstractRNG=Xoshiro(1234), showprogress::Bool=false, verbose::Bool=false, regularization::RT=-Inf) where {I<:Integer, RT<:Real, MT<:AbstractMatrix{RT}, D<:Distribution}

Run the Gaussian Expansion Cavity Method (GECaM) fixed-point algorithm for deterministic dynamics. It computes the fixed-point values of the mean abundances, correlations, and susceptibilities for a given instance of the disordered model.

Arguments

model::Model{Deterministic, I, RT, MT, D}: The model to run the GECaM on.

max_iter::I: Maximum number of iterations to run.

conv_threshold::RT: Convergence threshold for the fixed-point iterations.

damp::RT: Damping factor for the updates.

Keyword Arguments

init_type::Symbol: Initialization type for the nodes. Supported types are :zero, :random, and :custom (default is :random).

mu0::RT: Initial value for the mean (used if init_type is :custom, default is 0.0).

q0::RT: Initial value for the correlation (used if init_type is :custom, default is 0.0).

chi0::RT: Initial value for the susceptibility (used if init_type is :custom, default is 0.0).

rng::AbstractRNG: Random number generator for random initialization (default is Xoshiro(1234)).

showprogress::Bool: Whether to show progress during the iterations (default is false).

verbose::Bool: Whether to print detailed information during the iterations (default is false).

divergence_threshold::RT: Threshold for detecting divergence in the algorithm (default is 1e6).

regularization::RT: Regularization parameter to avoid negative values (default is -Inf).

Output

nodes::Vector{NodeFP{Deterministic, I, RT}}: The nodes with their updated cavity and marginal values after running the GECaM fixed-point algorithm.

converged::Bool: Whether the algorithm converged within the specified number of iterations.

diverged::Bool: Whether the algorithm diverged during the iterations.

RandomLotkaVolterraCavity.run_GECaM_FP — Method

run_GECaM_FP(model::ModelDisordered{Deterministic, I, RT, D1, D2, D3, FT}, P::I, max_iter::I, conv_threshold::RT, damp::RT; init_type::Symbol=:random, mu0::RT=0.0, q0::RT=0.0, chi0::RT=0.0, rng::AbstractRNG=Xoshiro(1234), showprogress::Bool=false, verbose::Bool=false, divergence_threshold::RT=1e6, regularization::RT=-Inf) where {I<:Integer, RT<:Real, D1<:Distribution, D2<:Distribution, D3<:Distribution, FT<:Function}

Run the Gaussian Expansion Cavity Method (GECaM) fixed-point algorithm for a disordered model with deterministic dynamics. It computes through a population dynamics algorithm the fixed-point values of the mean abundances, correlations, and susceptibilities averaged over the disordered model.

Arguments

model::ModelDisordered{Deterministic, I, RT, D1, D2, D3, FT}: The disordered model to run the GECaM on.

P::I: Number of elements of the populations.

max_iter::I: Maximum number of iterations to run.

conv_threshold::RT: Convergence threshold for the algorithm.

damp::RT: Damping factor for the updates.

Keyword Arguments

init_type::Symbol: Initialization type for the nodes. Supported types are :zero, :random, and :custom (default is :random).

mu0::RT: Initial value for the mean (used if init_type is :custom, default is 0.0).

q0::RT: Initial value for the correlation (used if init_type is :custom, default is 0.0).

chi0::RT: Initial value for the susceptibility (used if init_type is :custom, default is 0.0).

rng::AbstractRNG: Random number generator for random initialization (default is Xoshiro(1234)).

showprogress::Bool: Whether to show progress during the iterations (default is false).

verbose::Bool: Whether to print detailed information during the iterations (default is false).

divergence_threshold::RT: Threshold for detecting divergence in the algorithm (default is 1e6).

regularization::RT: Regularization parameter to avoid negative values (default is -Inf).

Output

cav_pop::PopFP{Deterministic, I, RT}: The cavity population with updated values after running the GECaM fixed-point algorithm.

marg_pop::PopFP{Deterministic, I, RT}: The marginal population with updated values after running the GECaM fixed-point algorithm.

converged::Bool: Whether the algorithm converged within the specified number of iterations.

diverged::Bool: Whether the algorithm diverged during the iterations.

RandomLotkaVolterraCavity.run_GECaM_FP_q0 — Method

run_GECaM_FP_q0(model::ModelDisordered{Deterministic, I, RT, D1, D2, D3, FT}, P::I, max_iter::I, conv_threshold::RT, damp::RT; init_type::Symbol=:random, x0::RT=0.0, chi0::RT=0.0, rng::AbstractRNG=Xoshiro(1234), showprogress::Bool=false, verbose::Bool=false, divergence_threshold::RT=1e6, regularization::RT=-Inf) where {I<:Integer, RT<:Real, D1<:Distribution, D2<:Distribution, D3<:Distribution, FT<:Function}

Run the Gaussian Expansion Cavity Method (GECaM) fixed-point algorithm for a disordered model with deterministic dynamics and simplified case where correlations are set to zero (q = 0). It computes through a population dynamics algorithm the fixed-point values of the mean abundances and susceptibilities averaged over the disordered model.

Arguments

model::ModelDisordered{Deterministic, I, RT, D1, D2, D3, FT}: The disordered model to run the GECaM on.

P::I: Number of elements of the populations.

max_iter::I: Maximum number of iterations to run.

conv_threshold::RT: Convergence threshold for the algorithm.

damp::RT: Damping factor for the updates.

Keyword Arguments

init_type::Symbol: Initialization type for the nodes. Supported types are :zero, :random, and :custom (default is :random).

x0::RT: Initial value for the mean (used if init_type is :custom, default is 0.0).

chi0::RT: Initial value for the susceptibility (used if init_type is :custom, default is 0.0).

rng::AbstractRNG: Random number generator for random initialization (default is Xoshiro(1234)).

showprogress::Bool: Whether to show progress during the iterations (default is false).

verbose::Bool: Whether to print detailed information during the iterations (default is false).

divergence_threshold::RT: Threshold for detecting divergence in the algorithm (default is 1e6).

regularization::RT: Regularization parameter to avoid negative values (default is -Inf).

Output

cav_pop::PopFP_q0{I, RT}: The cavity population with updated values after running the GECaM fixed-point algorithm.

marg_pop::PopFP_q0{I, RT}: The marginal population with updated values after running the GECaM fixed-point algorithm.

converged::Bool: Whether the algorithm converged within the specified number of iterations.

diverged::Bool: Whether the algorithm diverged during the iterations.